Start splitting out HttpClient WIP
