+++ core-android/src/test/java/com/mux/core_android/testdoubles/FakeNetwork.kt
@@ -3,7 +3,7 @@ package com.mux.core_android.testdoubles
 import android.net.Uri
 import com.mux.core_android.log
 import com.mux.stats.sdk.muxstats.INetworkRequest
-import com.mux.stats.sdk.muxstats.beaconAuthority
+import com.mux.android.http.beaconAuthority
 import com.mux.stats.sdk.muxstats.util.logTag
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
